network draft

+++ point.h
@@ -1,9 +1,11 @@
 #ifndef POINT_H
 #define POINT_H
 
+#include <stdint.h>
+
 typedef struct _Point {
-    int x;
-    int y;
+    int8_t x;
+    int8_t y;
 } Point;
 
 #endif
