+++ snake.h
@@ -1,6 +1,8 @@
 #ifndef SNAKE_H
 #define SNAKE_H
 
+#include <stdint.h>
+
 #include "defines.h"
 #include "point.h"
 
@@ -8,8 +10,8 @@ typedef Point SnakeSegment;
 
 typedef struct _Snake{
     SnakeSegment body[SNAKE_MAX_LENGTH];
-    int length; 
-    int speed;
+    uint8_t length; 
+    uint8_t  speed;
     Point direction;
 } Snake;
